Added script for keeping a Minecraft server up
This script keeps a Minecraft server (or whatever program you want) running in an endless loop. It has a few extra features like a wait time before restarting, and a message that says that the process has died. Don't use this for a program that forks off into the background after running the command, or you will likely have the process duplicating itself every 30 seconds (or whatver interval you set)!